tests/wpt/tests/css/css-values/calc-size/animation/interpolate-size-min-width-interpolation.html

<!DOCTYPE html>
<meta charset="UTF-8">
<title>min-width interpolation with interpolate-size: allow-keywords</title>
<link rel="help" href="https://drafts.csswg.org/css-sizing-3/#propdef-min-width">
<link rel="help" href="https://drafts.csswg.org/css-values-5/#interpolate-size">

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/interpolation-testcommon.js"></script>

<style>
:root {
  interpolate-size: allow-keywords;
}
.parent {
  min-width: 30px;
}
.target {
  width: 0px;
  height: 10px;
  min-width: 10px;
}
.target::before {
  display: block;
  content: "";
  width: 100px;
  height: 50px;
}
</style>

<body></body>

<script>

test_interpolation({
  property: 'min-width',
  from: 'initial',
  to: '20px',
}, [
  {at: -0.3, expect: 'calc-size(auto, -6px + size * 1.3)'},
  {at: 0, expect: 'calc-size(auto, 0px + size * 1)'},
  {at: 0.3, expect: 'calc-size(auto, 6px + size * 0.7)'},
  {at: 0.6, expect: 'calc-size(auto, 12px + size * 0.4)'},
  {at: 1, expect: 'calc-size(auto, 20px + size * 0)'},
  {at: 1.5, expect: 'calc-size(auto, 30px + size * -0.5)'}
]);

test_interpolation({
  property: 'min-width',
  from: 'unset',
  to: '20px',
}, [
  {at: -0.3, expect: 'calc-size(auto, -6px + size * 1.3)'},
  {at: 0, expect: 'calc-size(auto, 0px + size * 1)'},
  {at: 0.3, expect: 'calc-size(auto, 6px + size * 0.7)'},
  {at: 0.6, expect: 'calc-size(auto, 12px + size * 0.4)'},
  {at: 1, expect: 'calc-size(auto, 20px + size * 0)'},
  {at: 1.5, expect: 'calc-size(auto, 30px + size * -0.5)'}
]);

test_interpolation({
  property: 'min-width',
  from: 'auto',
  to: '20px',
}, [
  {at: -0.3, expect: 'calc-size(auto, -6px + size * 1.3)'},
  {at: 0, expect: 'calc-size(auto, 0px + size * 1)'},
  {at: 0.3, expect: 'calc-size(auto, 6px + size * 0.7)'},
  {at: 0.6, expect: 'calc-size(auto, 12px + size * 0.4)'},
  {at: 1, expect: 'calc-size(auto, 20px + size * 0)'},
  {at: 1.5, expect: 'calc-size(auto, 30px + size * -0.5)'}
]);

test_interpolation({
  property: 'min-width',
  from: '0px',
  to: 'stretch',
}, [
  {at: -0.3, expect: 'calc-size(stretch, 0px + size * -0.3)'},
  {at: 0, expect: 'calc-size(stretch, 0px + size * 0)'},
  {at: 0.3, expect: 'calc-size(stretch, 0px + size * 0.3)'},
  {at: 0.6, expect: 'calc-size(stretch, 0px + size * 0.6)'},
  {at: 1, expect: 'calc-size(stretch, 0px + size * 1)'},
  {at: 1.5, expect: 'calc-size(stretch, 0px + size * 1.5)'}
]);

test_no_interpolation({
  property: 'min-width',
  from: 'min-content',
  to: 'fit-content',
});

test_no_interpolation({
  property: 'min-width',
  from: 'auto',
  to: 'max-content',
});

test_interpolation({
  property: 'min-width',
  from: 'fit-content',
  to: neutralKeyframe, /* 10px */
}, [
  {at: -0.3, expect: 'calc-size(fit-content, -3px + size * 1.3)'},
  {at: 0, expect: 'calc-size(fit-content, 0px + size * 1)'},
  {at: 0.3, expect: 'calc-size(fit-content, 3px + size * 0.7)'},
  {at: 0.6, expect: 'calc-size(fit-content, 6px + size * 0.4)'},
  {at: 1, expect: 'calc-size(fit-content, 10px + size * 0)'},
  {at: 1.5, expect: 'calc-size(fit-content, 15px + size * -0.5)'}
]);

test_interpolation({
  property: 'min-width',
  from: neutralKeyframe, /* 10px */
  to: 'max-content',
}, [
  {at: -0.3, expect: 'calc-size(max-content, 13px + size * -0.3)'},
  {at: 0, expect: 'calc-size(max-content, 10px + size * 0)'},
  {at: 0.3, expect: 'calc-size(max-content, 7px + size * 0.3)'},
  {at: 0.6, expect: 'calc-size(max-content, 4px + size * 0.6)'},
  {at: 1, expect: 'calc-size(max-content, 0px + size * 1)'},
  {at: 1.5, expect: 'calc-size(max-content, -5px + size * 1.5)'}
]);

</script>
