third_party/blink/web_tests/external/wpt/css/css-values/calc-size/animation/interpolate-size-min-width-interpolation-expected.txt

This is a testharness.js-based test.
Found 24 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (-0.3) should be [calc-size(stretch, 0px + size * -0.3)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (0) should be [calc-size(stretch, 0px + size * 0)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (0.3) should be [calc-size(stretch, 0px + size * 0.3)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (0.6) should be [calc-size(stretch, 0px + size * 0.6)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (1) should be [calc-size(stretch, 0px + size * 1)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions: property <min-width> from [0px] to [stretch] at (1.5) should be [calc-size(stretch, 0px + size * 1.5)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (-0.3) should be [calc-size(stretch, 0px + size * -0.3)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (0) should be [calc-size(stretch, 0px + size * 0)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (0.3) should be [calc-size(stretch, 0px + size * 0.3)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (0.6) should be [calc-size(stretch, 0px + size * 0.6)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (1) should be [calc-size(stretch, 0px + size * 1)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Transitions with transition: all: property <min-width> from [0px] to [stretch] at (1.5) should be [calc-size(stretch, 0px + size * 1.5)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (-0.3) should be [calc-size(stretch, 0px + size * -0.3)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (0) should be [calc-size(stretch, 0px + size * 0)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (0.3) should be [calc-size(stretch, 0px + size * 0.3)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (0.6) should be [calc-size(stretch, 0px + size * 0.6)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (1) should be [calc-size(stretch, 0px + size * 1)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] CSS Animations: property <min-width> from [0px] to [stretch] at (1.5) should be [calc-size(stretch, 0px + size * 1.5)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (-0.3) should be [calc-size(stretch, 0px + size * -0.3)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (0) should be [calc-size(stretch, 0px + size * 0)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (0.3) should be [calc-size(stretch, 0px + size * 0.3)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (0.6) should be [calc-size(stretch, 0px + size * 0.6)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (1) should be [calc-size(stretch, 0px + size * 1)]
  assert_true: 'to' value should be supported expected true got false
[FAIL] Web Animations: property <min-width> from [0px] to [stretch] at (1.5) should be [calc-size(stretch, 0px + size * 1.5)]
  assert_true: 'to' value should be supported expected true got false
Harness: the test ran to completion.
